Fix build of xf86vidmode
This just removes an optimisation and introduces useless calls to
From::from(). However, xf86vidmode has one case where this actually
makes a difference.

In the future, this optimisation might be re-introduced. For now I am
lazy and just want things to compile again.

Build failure was:

error[E0369]: no implementation for `ClockFlag & u32`
    --> x11rb-protocol\src\protocol\xf86vidmode.rs:2341:82
     |
2341 | ...   assert_eq!(self.clock.len(), usize::try_from(1u32.checked_sub(self.flags & 1u32).unwrap().checked_mul(self.clocks).unwrap()).unwrap...
     |                                                                     ---------- ^ ---- u32
     |                                                                     |
     |                                                                     ClockFlag
     |
note: an implementation of `BitAnd<_>` might be missing for `ClockFlag`
    --> x11rb-protocol\src\protocol\xf86vidmode.rs:111:1
     |
111  | pub struct ClockFlag(u32);
     | ^^^^^^^^^^^^^^^^^^^^^^^^^^ must implement `BitAnd<_>`
note: the following trait must be implemented
error[E0308]: mismatched types

+++ x11rb-protocol/src/protocol/dri2.rs
@@ -513,7 +513,7 @@ impl TryParse for ConnectReply {
         let remaining = remaining.get(16..).ok_or(ParseError::InsufficientData)?;
         let (driver_name, remaining) = crate::x11_utils::parse_u8_list(remaining, driver_name_length.try_to_usize()?)?;
         let driver_name = driver_name.to_vec();
-        let (alignment_pad, remaining) = crate::x11_utils::parse_u8_list(remaining, (driver_name_length.checked_add(3u32).ok_or(ParseError::InvalidExpression)? & (!3u32)).checked_sub(driver_name_length).ok_or(ParseError::InvalidExpression)?.try_to_usize()?)?;
+        let (alignment_pad, remaining) = crate::x11_utils::parse_u8_list(remaining, (u32::from(driver_name_length).checked_add(3u32).ok_or(ParseError::InvalidExpression)? & (!3u32)).checked_sub(u32::from(driver_name_length)).ok_or(ParseError::InvalidExpression)?.try_to_usize()?)?;
         let alignment_pad = alignment_pad.to_vec();
         let (device_name, remaining) = crate::x11_utils::parse_u8_list(remaining, device_name_length.try_to_usize()?)?;
         let device_name = device_name.to_vec();
@@ -547,7 +547,7 @@ impl Serialize for ConnectReply {
         device_name_length.serialize_into(bytes);
         bytes.extend_from_slice(&[0; 16]);
         bytes.extend_from_slice(&self.driver_name);
-        assert_eq!(self.alignment_pad.len(), usize::try_from((driver_name_length.checked_add(3u32).unwrap() & (!3u32)).checked_sub(driver_name_length).unwrap()).unwrap(), "`alignment_pad` has an incorrect length");
+        assert_eq!(self.alignment_pad.len(), usize::try_from((u32::from(driver_name_length).checked_add(3u32).unwrap() & (!3u32)).checked_sub(u32::from(driver_name_length)).unwrap()).unwrap(), "`alignment_pad` has an incorrect length");
         bytes.extend_from_slice(&self.alignment_pad);
         bytes.extend_from_slice(&self.device_name);
     }
