+++ x11rb-protocol/src/protocol/sync.rs
@@ -1162,7 +1162,7 @@ pub struct CreateAlarmAux {
 }
 impl CreateAlarmAux {
     fn try_parse(value: &[u8], value_mask: u32) -> Result<(Self, &[u8]), ParseError> {
-        let switch_expr = value_mask;
+        let switch_expr = u32::from(value_mask);
         let mut outer_remaining = value;
         let counter = if switch_expr & u32::from(CA::COUNTER) != 0 {
             let remaining = outer_remaining;
@@ -1401,7 +1401,7 @@ pub struct ChangeAlarmAux {
 }
 impl ChangeAlarmAux {
     fn try_parse(value: &[u8], value_mask: u32) -> Result<(Self, &[u8]), ParseError> {
-        let switch_expr = value_mask;
+        let switch_expr = u32::from(value_mask);
         let mut outer_remaining = value;
         let counter = if switch_expr & u32::from(CA::COUNTER) != 0 {
             let remaining = outer_remaining;
